requirements.txt
Flask
Werkzeug
python-docx
pdfplumber
scikit-learn
